refactor: Move functions around and ingest system prompt in all bench related prompts
